requirements.txt
# BeliefBank rule generation (LLM pipeline)
openai
LTNtorch
python-dotenv
